requirements.txt
requests
beautifulsoup4
selenium
langchain
faiss-cpu
numpy
streamlit
python-dotenv
google-generativeai
configparser
